backend/pyproject.toml
[project]
name = "astitva"
version = "0.1.0"
description = "A Intelligent Attendance System"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "firebase>=4.0.1",
    "firebase-admin>=7.1.0",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
    "bcrypt>=5.0.0",
    "jwt>=1.4.0",
    "sqlalchemy>=2.0.45",
    "pillow>=12.1.0",
    "opencv-python>=4.11.0.86",
    "python-multipart>=0.0.21",
    "python-jose>=3.5.0",
    "email-validator>=2.3.0",
]
